refactor bg-main.js
