+++ source/js/buyers-guide/index.js
@@ -1,0 +1,5 @@
+export {
+  bindWindowEventHandlers,
+  bindEventHandlers,
+} from "./template-js-handler";
+export { injectReactComponents } from "./inject-react";
