+++ source/js/buyers-guide/template-js-handler/index.js
@@ -1,0 +1,8 @@
+import copyLinkHandler from "./copy-link.js";
+
+/**
+ * Bind event handlers
+ */
+export const bindEventHandlers = () => {
+  copyLinkHandler();
+};
